Add make-legacy-parameter
It turned out that make-thread-parameter as defined in SRFI-226
is not compatible with our legacy parameter.  So we add make-legacy-parameter
for the compatibility.
See https://github.com/shirok/Gauche/issues/935 for the details.

NB: Current make-thread-parameter does not precisely work as SRFI-226,
but just the same as legacy parameters.  We'll address this later.