Unary + and * now invokes object-{+,*} on non-numeric arg
If you want to extend arithmetic operators for your class, you should
now define unary object-+ and object-* as well.  They should just return
the argument as is; the existence of the method vouches it's a valid
operation.

For the backward compatibility, we return the argument as is if
there's no specialized unary method defined, but with a warning.

https://github.com/shirok/Gauche/issues/1012