gauche.config - Show original path when replace-install-dir fails
`gauche-config` replaces '@' in paths with the intallation directory.
Currently it is supposed to be used only on MinGW and OSX-Framework
builds.  However, when we make some mistake in configure and some
output parameters are not substitued, it can cause unexpected errors.
Showing the original string in the error message will help to track
down the cause.