peg - Fix $expect on returning failed stream position
Previously it returned the beginning position when match fails.
It effectively makes $expect works as if $try; failure "rewinds" the
input so that alternative branch may be tried.

It is debatable how $expect should work in that case, but it seems
more reasonable that $expect only modifies failure messages and leave
other properties of the given parser intact.  If the user wants to
rewind to try next alternative, they can always use $try explicitly.