Let cf-path-x return results instead of setting subst params
In autoconf, AC_PATH_X sets a few shell variables. In Scheme it is
awkward to communicate via global variables.  So initially we set
substitution params, but that's awkward, too, because they won't
really used for substition.  So, after some consideration,
we decided that cf-path-x return the check result as multiple
values, for that's how Schemers do.