Move away from ep->prev chain when UNIFY_ERROR_HANDLING
Another step towards https://github.com/shirok/Gauche/issues/1037

Previouly, ep->prev chain has two roles: It is a stack of erorr handler
chain, and it is also a list of escapePoints that may contain a
pointer to in-stack cont frames (thus needs to be updated when
continuation frames are moved to heap).

With UNIFY_ERROR_HANDLING, error handler stack is unified to exception
handlers which is in denv, so we only need to keep track of EPs for
continuation frame moving.  We keep them in vm->floatingEscapePoints
list (which is a Scheme list).  Don't confuse this from 'floating' pointer
we used to have in EP; that had a different role.