Avoid infinite recursion to %pretty-print
`pprint` calls built-in writer for non-aggregate objects to obtain
rendered output.  The built-in writer used to call '%pretty-print' in
pass2 if :print-pretty is true, that caused infinite loop.  So we had
a check to avoid it in the pputil side. However, we don't need it
if the built-in writer knows when it should call '%pretty-print'.

Related: https://github.com/shirok/Gauche/issues/1058

Note: If this version of pputil is used with older versions of Gauche,
it goes to infinite recursion.  So gauche.pputil is added to preload.