Experimentally add make-id-transformer
Using keyword arg to er-macro-transformer is awkward, so we added
'wrapper' style API to create an id transformer.  This approach creates
an extra macro transformer during compilation, but that's just one-time
thing and doesn't matter much.  The actual macro expansion has no overhead.

The advantage of the wrapper style is that it can be used for different
transformers.

https://github.com/shirok/Gauche/issues/1061

`make-id-transformer` is similar to R6RS `make-variable-transformer`,
but we need `make-id-transformer` even you don't need `set!` expansion.