Fix test from changes of sys-exec
Now that we search an executable by ourselves rather than relying on
the system, the behavior of sys-exec is slightly changed on Windows.
Windows unconditionally searches the current directly, while we only
search the current directory if "." is in PATH.

This is compatibility-breaking change; however, we think Windows is
broken in this regard.  For now, we fix the test.  See how it goes.