Let number reader save error message
If the number reader is called with numread_packet->throwerror is FALSE,
the internal reader just returns #f on parse error and the detailed
error msg was lost.  Now we keep the error message in numread_packet
so that the public API can pass it to caller.

Preparation to fix https://github.com/shirok/Gauche/issues/912.