Replace *load-path* etc. with parameters
Although it isn't changed frequently, the effect of changing it should
be delimited, so parameters are more suitable.

To keep the backward compatibility with code dereferencing *load-path*
and *dynamic-load-path* directly, we provide autoloaded identifier macros.
Those are supposed to be changed with add-load-path, so there shouldn't
be code that set! to those variables.

We also changed *load-suffixes* and *load-path-hooks* to parmaeters.
These haven't been documented, but they are around for a long time,
there can be code that depends on them.  We'll take care of them
as we found them.