Fix precomp expansion of define-macro
Precomp used to expand define-macro to
   (define name (%make-macro-transformer ...))
This is technically wrong, since macro trasformer should be bonud with
define-syntax.  However, simply using define-syntax here won't cut it,
for we need a special treatment to keep the macro in the output code
(instead of fully consumed within the compiler).

Now we expand define-macro to the call to %insert-syntax-binding.
It needed some rearrangement of the code b/c initialization dependencies.